uacce: fix the big_memeory reserve
driver inclusion
category:bugfix
bugzilla:NA
CVE:NA

when u reserved a big memory(eg: 4G), (unsigned int)nr_pages is too big,
so "nr_pages << PAGE_SHIFT" overflows, the kernel may alloc the memory
to u, but the vm is wrong.

Feature or Bugfix: Bugfix

+++ include/linux/uacce.h
@@ -41,7 +41,7 @@ struct uacce_qfile_region {
 	enum uacce_qfrt type;
 	unsigned long iova;	/* iova share between user and device space */
 	struct page **pages;
-	unsigned int nr_pages;
+	unsigned long nr_pages;
 	int prot;
 	unsigned int flags;
 	struct list_head qs;	/* qs sharing the same region, for ss */
